mypy-requirements.txt
mypy==0.991
pandas-stubs

